Display likes number on DOM
