+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles/style.scss';
 import logo from './assets/logo.svg';
-import getLikes from './getlikes.js';
+import getLikes from './modules/getlikes.js';
 import { getMovies, API_URL, SEARCH_API } from './modules/getMovies.js';
 
 getMovies(API_URL);
